requirements.txt
openai>=1.0.0
aiohttp
pandas
yfinance
newspaper3k
serpapi
lxml_html_clean==0.4.2